services/ai/pyproject.toml
[tool.poetry]
name = "cliniqai-ai"
version = "0.0.1"
description = "CliniqAI AI Service — MedGemma-powered clinical decision support"
authors = ["CliniqAI Team"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = { version = "^0.29.0", extras = ["standard"] }
pydantic = "^2.7.1"
pydantic-settings = "^2.2.1"
google-generativeai = "^0.7.0"      # Gemini API (free tier, default)
google-cloud-aiplatform = "^1.51.0" # Vertex AI (MedGemma, optional)
google-auth = "^2.29.0"
python-multipart = "^0.0.9"
httpx = "^0.27.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.2"
pytest-asyncio = "^0.23.6"
ruff = "^0.4.3"
mypy = "^1.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
